Creating reusable classes to split up code
